refactor(validator): isolate DKG board upload protocol
